feat(xtask): allow a seed when generating consensus configs (#381)
This patch adds an option `--seed` to `cargo x generate-config`. The
seed is for generating a deterministic, not (pseudo-) random set of
signatures and bls12-381 shares. This is for use in tests and local
development to keep keep validator names and their configs stable.

This patch also removes the random selection of bootstrappers from the
previously randomly generated validators. There is no benefit for this.
Now the first n validators are selected as bootstrappers (with n set by
the `--bootstrappers` option).

Closes #378