chore: migrate allegretto docs to tempo repo (#1209)
After this PR the docs repo content should be fully in sync with the
tempo/docs folder.

The next step is to add the vercel build processes and relevant CIs. 

Note: The migration was done using `rsync` so it should be safe to
assume that the content was migrated correctly. But if you see any
discrepancies in changes you have made recently, please let me know.