feat: add `ReadOnlyStorageProvider` to use precompiles anywhere (#1546)
## Motivation

there are still several places in the codebase (like the revm handler,
or the tx-pool validator), where we perform manual storage slot checks
on the precompiles.

This wasn't initially tackled with the precompile refactor, but is a
potential source of human-errors, as we need to manually handle slots
and their encoding/decoding.

## Solution

add a new `struct ReadOnlyStorageProvider` which partially implements
`PrecompileStorageProvider` which panicks on all non-supported functions
(which are no-op and should never be called).

---------

Co-authored-by: Arsenii Kulikov <klkvrr@gmail.com>